Move ApiError interface to a more specific file
It's been moved to a more broad folder because it will be used by
multiple files.
