+++ client/data/capital/actions.ts
@@ -3,9 +3,9 @@
 /**
  * Internal Dependencies
  */
+import { ApiError } from '../../types/errors';
 import ACTION_TYPES from './action-types';
 import {
-	ApiError,
 	Summary,
 	LoansList,
 	UpdateSummaryAction,
