+++ client/data/capital/selectors.ts
@@ -3,7 +3,8 @@
 /**
  * Internal dependencies
  */
-import { ApiError, CapitalLoan, CapitalState, State, Summary } from './types';
+import { ApiError } from '../../types/errors';
+import { CapitalLoan, CapitalState, State, Summary } from './types';
 
 /**
  * Retrieves the Capital loans state from the wp.data store if the state
