+++ client/types/errors.d.ts
@@ -1,0 +1,3 @@
+export interface ApiError {
+	code: string;
+}
